+++ CHANGELOG.md
@@ -6,3 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/orisai/scheduler/compare/1.0.0...v2.x)
+
+### Changed
+
+- `JobManager`
+	- `getPair()` -> `getScheduledJob()`
+	- `getPairs()` -> `getScheduledJobs()`
+- `Scheduler`
+	- `getJobs()` -> `getScheduledJobs()`